+++ lib/src/inputs/bool/bool_field.dart
@@ -4,8 +4,8 @@
 
 // ignore_for_file: avoid_positional_boolean_parameters
 
-import 'package:flutter/material.dart';
-import 'package:flutter_input_library/flutter_input_library.dart';
+import "package:flutter/material.dart";
+import "package:flutter_input_library/flutter_input_library.dart";
 
 class BoolFormField extends FormField<bool> {
   BoolFormField({