+++ lib/src/inputs/text/password.dart
@@ -2,8 +2,8 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
+import "package:flutter/material.dart";
+import "package:flutter/services.dart";
 
 /// Generates a [TextFormField] for passwords. It requires a
 /// [FlutterFormInputController] as the [controller] parameter and an